semantic_sentence_comparer/requirements.txt
pandas
python-dotenv
openai
google-generativeai
transformers
torch
sentence-transformers
pytest
pytest-mock
